chore: extract transaction-search to new package
