+++ packages/suite/package.json
@@ -61,6 +61,7 @@
         "@suite-common/toast-notifications": "workspace:*",
         "@suite-common/token-definitions": "workspace:*",
         "@suite-common/trading": "workspace:*",
+        "@suite-common/transaction-search": "workspace:*",
         "@suite-common/tx-simulation": "workspace:^",
         "@suite-common/validators": "workspace:*",
         "@suite-common/wallet-config": "workspace:*",
