+++ suite-common/transaction-search/src/advancedSearchTransactions.ts
@@ -1,0 +1,54 @@
+import { AccountLabels } from '@suite-common/metadata-types';
+import { WalletAccountTransaction } from '@suite-common/wallet-types';
+
+import { simpleSearchTransactions } from './simpleSearchTransactions';
+
+export const advancedSearchTransactions = (
+    transactions: WalletAccountTransaction[],
+    accountMetadata: AccountLabels,
+    search: string,
+) => {
+    // No AND/OR operators, just run a simple search
+    if (!search.includes('&') && !search.includes('|')) {
+        return simpleSearchTransactions(transactions, accountMetadata, search);
+    }
+
+    // Split by OR operator first
+    let orSplit = search.split('|').filter(s => s.trim() !== '');
+    if (!orSplit || orSplit.length === 1) {
+        orSplit = [search.replace('|', '')];
+    }
+
+    // Get all TxIDs matching the searches
+    const filteredTxIDs = new Set([
+        ...orSplit.flatMap(or => {
+            // And searches (only keep results that appear X (split) times)
+            const andSplit = or.split('&');
+            if (!andSplit || andSplit.length === 1) {
+                return simpleSearchTransactions(
+                    transactions,
+                    accountMetadata,
+                    or.replace('&', ''),
+                ).flatMap(t => t.txid);
+            }
+
+            const andTxs = andSplit.flatMap(and =>
+                simpleSearchTransactions(transactions, accountMetadata, and).map(t => t.txid),
+            );
+
+            const transactionCount: { [txid: string]: number } = {};
+
+            return andTxs.filter(txid => {
+                if (!transactionCount[txid]) {
+                    transactionCount[txid] = 0;
+                }
+
+                transactionCount[txid]++;
+
+                return transactionCount[txid] === andSplit.length;
+            });
+        }),
+    ]);
+
+    return transactions.filter(t => filteredTxIDs.has(t.txid));
+};
