+++ suite-common/transaction-search/src/index.ts
@@ -1,0 +1,1 @@
+export { advancedSearchTransactions } from './advancedSearchTransactions';
