+++ suite-common/transaction-search/src/searchOperations.ts
@@ -1,0 +1,1 @@
+export const searchOperators = ['<', '>', '=', '!='] as const;
