+++ suite-common/transaction-search/tsconfig.json
@@ -1,0 +1,11 @@
+{
+    "extends": "../../tsconfig.base.json",
+    "compilerOptions": { "outDir": "libDev" },
+    "include": [".", "**/*.json"],
+    "references": [
+        { "path": "../metadata-types" },
+        { "path": "../wallet-types" },
+        { "path": "../wallet-utils" },
+        { "path": "../../packages/utils" }
+    ]
+}
