+++ suite-common/wallet-utils/src/transactionUtils.ts
@@ -1,7 +1,6 @@
 import { addDays, startOfMonth } from 'date-fns';
 import { fromWei, toWei } from 'web3-utils';
 
-import { AccountLabels } from '@suite-common/metadata-types';
 import { SignOperator } from '@suite-common/suite-types';
 import { NetworkType, getNetworkType } from '@suite-common/wallet-config';
 import {
@@ -38,7 +37,6 @@ import { isStakeTypeTx } from './ethereumStakingUtils';
 import { toFiatCurrency } from './fiatConverterUtils';
 import { getFiatRateKey, roundTimestampToNearestPastHour } from './fiatRatesUtils';
 import { getMyInputsFromTransaction } from './getMyInputsFromTransaction';
-import { isTokenTransferMatchesSearch } from './tokenUtils';
 
 export const sortByBlockHeight = (a: { blockHeight?: number }, b: { blockHeight?: number }) => {
     // if both are missing the blockHeight don't change their order
@@ -882,296 +880,6 @@ export const enhanceTransaction = (
     rbfParams: getRbfParams(origTx, account),
     hex: (origTx.blockHeight ?? 0) <= 0 && origTx.rbf ? origTx.hex : undefined, // store tx hex **only** for pending transactions (used by rbf)
 });
-
-const groupTransactionIdsByAddress = (transactions: WalletAccountTransaction[]) => {
-    const addresses: Record<string, string[]> = {};
-    const addAddress = (txid: string, addrs: string[] | undefined) => {
-        if (!addrs) {
-            return;
-        }
-
-        addrs.forEach(address => {
-            if (!addresses[address]) {
-                addresses[address] = [];
-            }
-
-            if (addresses[address].indexOf(txid) === -1) {
-                addresses[address].push(txid);
-            }
-        });
-    };
-
-    transactions.forEach(t => {
-        // Inputs
-        t.details.vin.forEach(vin => addAddress(t.txid, vin.addresses));
-        // Outputs
-        t.details.vout.forEach(vout => addAddress(t.txid, vout.addresses));
-        // Targets
-        t.targets.forEach(target => addAddress(t.txid, target.addresses));
-    });
-
-    return addresses;
-};
-
-const groupTransactionsByLabel = (accountMetadata: AccountLabels) => {
-    const labels: Record<string, string[]> = {};
-    const { outputLabels } = accountMetadata;
-
-    typedObjectKeys(outputLabels).forEach(txid => {
-        Object.values(outputLabels[txid]).forEach(label => {
-            if (!labels[label]) {
-                labels[label] = [];
-            }
-
-            labels[label].push(txid);
-        });
-    });
-
-    return labels;
-};
-
-const groupAddressesByLabel = (accountMetadata: AccountLabels) => {
-    const labels: Record<string, string[]> = {};
-    const { addressLabels } = accountMetadata;
-
-    typedObjectKeys(addressLabels).forEach(address => {
-        const label = addressLabels[address];
-
-        if (!labels[label]) {
-            labels[label] = [];
-        }
-
-        labels[label].push(address);
-    });
-
-    return labels;
-};
-
-const getTargetAmounts = (transaction: WalletAccountTransaction) =>
-    transaction.targets.length === 0
-        ? [formatNetworkAmount(transaction.amount, transaction.symbol)]
-        : transaction.targets.flatMap(target => getTargetAmount(target, transaction) || []);
-
-const searchOperators = ['<', '>', '=', '!='] as const;
-const numberSearchFilter = (
-    transaction: WalletAccountTransaction,
-    amount: BigNumber,
-    operator: (typeof searchOperators)[number],
-) => {
-    const targetAmounts = getTargetAmounts(transaction);
-    const op = getTxOperation(transaction.type);
-    if (!op) {
-        return false;
-    }
-
-    return (
-        targetAmounts.filter(targetAmount => {
-            let bnTargetAmount = new BigNumber(targetAmount);
-            if (op === 'negative') {
-                bnTargetAmount = bnTargetAmount.negated();
-            }
-
-            switch (operator) {
-                case '<':
-                    return bnTargetAmount.lte(amount);
-                case '>':
-                    return bnTargetAmount.gte(amount);
-                case '=':
-                    return bnTargetAmount.eq(amount);
-                case '!=':
-                    return !bnTargetAmount.eq(amount);
-                default:
-                    return false;
-            }
-        }).length > 0
-    );
-};
-
-const searchDateRegex = new RegExp(/^[0-9]{4}-[0-9]{2}-[0-9]{2}$/);
-export const simpleSearchTransactions = (
-    transactions: WalletAccountTransaction[],
-    accountMetadata: AccountLabels, // Todo: this is wrong, this shall reflect SuiteSync labels. See https://github.com/trezor/trezor-suite/issues/24803
-    search: string,
-) => {
-    // Trim
-    search = search.trim();
-
-    // If the string is empty or only contains search operators, there's no search
-    if (['', ...searchOperators].includes(search)) {
-        return transactions;
-    }
-
-    // Check for date
-    if (searchDateRegex.test(search)) {
-        // Add search operator so it gets picked up below
-        search = `=${search}`;
-    }
-
-    // If it's an amount search (starting with <, > or = operator)
-    const searchOperator = searchOperators.find(k => search.startsWith(k));
-    if (searchOperator) {
-        // Remove search operator from search string
-        search = search.replace(searchOperator, '').trim();
-
-        // Is date?
-        if (searchDateRegex.test(search)) {
-            const timestamp = +new Date(`${search}T00:00:00Z`) / 1000;
-            switch (searchOperator) {
-                case '>':
-                    return transactions.filter(t => t.blockTime && t.blockTime > timestamp);
-                case '<':
-                    return transactions.filter(
-                        t => t.blockTime && t.blockTime < timestamp + 24 * 60 * 60,
-                    );
-                case '=':
-                    return transactions.filter(
-                        t =>
-                            t.blockTime &&
-                            t.blockTime > timestamp &&
-                            t.blockTime < timestamp + 24 * 60 * 60,
-                    );
-                case '!=':
-                    return transactions.filter(
-                        t =>
-                            t.blockTime &&
-                            (t.blockTime < timestamp || t.blockTime > timestamp + 24 * 60 * 60),
-                    );
-                // no default
-            }
-        }
-
-        // Is number?
-        if (!Number.isNaN(search)) {
-            const amount = new BigNumber(search);
-
-            return transactions.filter(t => numberSearchFilter(t, amount, searchOperator));
-        }
-
-        return [];
-    }
-
-    const txsToSearch: string[] = [];
-
-    // Searching for an amount (without operator)
-    if (!Number.isNaN(search)) {
-        const foundTxsForNumber = transactions.flatMap(transaction => {
-            const targetAmounts = getTargetAmounts(transaction);
-            if (targetAmounts.filter(targetAmount => targetAmount.includes(search)).length === 0) {
-                return [];
-            }
-
-            return transaction.txid;
-        });
-        txsToSearch.push(...foundTxsForNumber);
-    }
-
-    // Find by output label
-    const txsForOutputLabels = groupTransactionsByLabel(accountMetadata);
-    const foundTxsForOutputLabel = typedObjectKeys(txsForOutputLabels).flatMap(label => {
-        if (label.toLowerCase().includes(search.toLowerCase())) {
-            return txsForOutputLabels[label];
-        }
-
-        return [];
-    });
-    txsToSearch.push(...foundTxsForOutputLabel);
-
-    // Find by address label
-    const addressesForLabel = groupAddressesByLabel(accountMetadata);
-    const foundAddressesForLabel = typedObjectKeys(addressesForLabel).flatMap(label => {
-        if (label.toLowerCase().includes(search.toLowerCase())) {
-            return addressesForLabel[label];
-        }
-
-        return [];
-    });
-
-    // Find by address
-    const txsForAddresses = groupTransactionIdsByAddress(transactions);
-    const foundTxsForAddress = typedObjectKeys(txsForAddresses).flatMap(address => {
-        if (
-            address.toLowerCase().includes(search.toLowerCase()) ||
-            foundAddressesForLabel.includes(address)
-        ) {
-            return txsForAddresses[address];
-        }
-
-        return [];
-    });
-    txsToSearch.push(...foundTxsForAddress);
-
-    // Find by token name, symbol or contract
-    const foundTxsForToken = transactions.flatMap(transaction => {
-        const hasMatchingToken = transaction.tokens.some(
-            token =>
-                isTokenTransferMatchesSearch(token, search.toLowerCase()) ||
-                token.to?.toLowerCase().includes(search.toLowerCase()) ||
-                token.from?.toLowerCase().includes(search.toLowerCase()),
-        );
-
-        if (hasMatchingToken) {
-            return transaction.txid;
-        }
-
-        return [];
-    });
-    txsToSearch.push(...foundTxsForToken);
-
-    // Remove duplicate txIDs
-    return transactions.filter(
-        t => [...new Set(txsToSearch)].includes(t.txid) || t.txid.includes(search),
-    );
-};
-
-export const advancedSearchTransactions = (
-    transactions: WalletAccountTransaction[],
-    accountMetadata: AccountLabels,
-    search: string,
-) => {
-    // No AND/OR operators, just run a simple search
-    if (!search.includes('&') && !search.includes('|')) {
-        return simpleSearchTransactions(transactions, accountMetadata, search);
-    }
-
-    // Split by OR operator first
-    let orSplit = search.split('|').filter(s => s.trim() !== '');
-    if (!orSplit || orSplit.length === 1) {
-        orSplit = [search.replace('|', '')];
-    }
-
-    // Get all TxIDs matching the searches
-    const filteredTxIDs = new Set([
-        ...orSplit.flatMap(or => {
-            // And searches (only keep results that appear X (split) times)
-            const andSplit = or.split('&');
-            if (!andSplit || andSplit.length === 1) {
-                return simpleSearchTransactions(
-                    transactions,
-                    accountMetadata,
-                    or.replace('&', ''),
-                ).flatMap(t => t.txid);
-            }
-
-            const andTxs = andSplit.flatMap(and =>
-                simpleSearchTransactions(transactions, accountMetadata, and).map(t => t.txid),
-            );
-
-            const transactionCount: { [txid: string]: number } = {};
-
-            return andTxs.filter(txid => {
-                if (!transactionCount[txid]) {
-                    transactionCount[txid] = 0;
-                }
-
-                transactionCount[txid]++;
-
-                return transactionCount[txid] === andSplit.length;
-            });
-        }),
-    ]);
-
-    return transactions.filter(t => filteredTxIDs.has(t.txid));
-};
 
 export const getTxHeaderSymbol = (transaction: WalletAccountTransaction) => {
     const isSingleTokenTransaction = transaction.tokens.length === 1;
